add navigation footnote component
